Add sff (San Francisco Faucet)
This commit adds many files inside sff directory.

The important ones are:

 - simplereplnn.sh
 - sff-loop.sh
 - blocknotify-signet.sh
 - walletnotify-signet.sh
 - refresh.sh

The other files are their dependencies and are meant to be
reachable within PATH.

All could be generalized and simplified for use with any testnet
or custom signet network, but this is the current state and it seems
to [work](https://mempool.space/signet/address/tb1p4tp4l6glyr2gs94neqcpr5gha7344nfyznfkc8szkreflscsdkgqsdent4)

RBF replacements may rather use weight for computations and the "stage 4"
of `simplereplnn.sh` could be eliminated. But that is for future play.
The point of this commit is to preserve it as it is now.

The two main wallets are refreshed from their `listdescriptors` dumps.
See also https://github.com/bitcoin/bitcoin/issues/31816 for an idea how
to improve Bitcoin Core wallet for faucets.

`bch.sh` is for "`bitcoin-cli` here", i.e. a helper script that cares
about from where it is called ($PWD). Possible impovement would be an
environment variable with a directory where to operate.

Basically there are three stages: sff, sff-s2 and sff-s3.
`signetwallet.sh` makes files in /tmp/sff and get moved next in turns.
It was my idea how to cope with leftovers if block comes too quickly
and some addresses were left behind. They shall be paid in next block.

Compared to the previous mode of operation the whole faucet now operates
in asynchronous time for the user. The requests are saved into temporary
files which are regularly read by the `sff-loop.sh` script.

Fixes #4