sff: Sync current runnig state using V3
There will soon be a change to V2 end-of-ancestor-chain.
Think about it in the meanwhile. Something like
155da10f444d3650f14a8411bfe9ade76131406f58ad97434f8501ec917bb851
and its 24 ancestors in block 237342 (and do not care about its
descendant that was mined merely in block 237343).

The replacements will be done like now but just at the end of
this 25-together in-mempool chain of transactions. All the nodes
running with the default setting of -limitdescendantcount=25 will
not care about any fee-changing 26th transaction which a faucet
payout receiver may send to the network. Some food for thought.